make consistent with single and double floating point precision

+++ modules/aruco/include/opencv2/charuco.hpp
@@ -103,8 +103,8 @@ public:
      * This functions creates a CharucoBoard object given the number of squares in each direction
      * and the size of the markers and chessboard squares.
      */
-    CV_EXPORTS static CharucoBoard create(int squaresX, int squaresY, float squareLength,
-                                          float markerLength, DICTIONARY dictionary);
+    CV_EXPORTS static CharucoBoard create(int squaresX, int squaresY, double squareLength,
+                                          double markerLength, DICTIONARY dictionary);
 
     /**
       *
